Trac #80: attributes treated incorrectly with multiple levels of restriction/extension
PyXB generates unique names for each attribute declaration.  An attribute
use, which is what appears in the generated class, may be a restriction of
another attribute use, in which case the generated names should correspond
to the names used in that use.  The code improperly traversed only a single
restrictionOf step when determining the names to be used.