trac/104 error resolving circular dependencies
Due to cycles in the component dependency graph among namespaces, the
example schema in this ticket could improperly fail to resolve, depending on
the order in which the namespaces were attempted.  Augment the unresolved
component infrastructure to retain references to components on which
resolution depends; resolve independent sibling namespaces first, and allow
for the possibility that multiple iterations through the set will be
required before all dependencies can be resolved.