trac/116: Simple-typed elements accept any attributes
Move the instance attribute initialization code from complexTypeDefinition
to _TypeBinding_mixin, so that the attempt to set attributes from a DOM
source is made visible.  Add support in simpleTypeDefinition to record the
attempt to set an attribute: this raises the same exception as is raised in
the complexTypeDefinition version when the attribute is not recognized.

The comments in the code now imply that attribute settings are simply
accepted, and a validation error occurs only when requested.  As noted in
trac/117 complex types have always raised exceptions immediately on an
attempt to construct with an unrecognized attribute, regardless of
validation settings.  This may or may not be a bug (more thought is required
on when the diagnostic for setting an attribute should be generated), but
it's not related to the issue in trac/116, so it's being deferred.