trac/119: unable to resolve with absent namespace
As absent namespaces cannot be identified within the document, the only
way to convert a document using an absent namespace to a Python instance  is
to do so through the CreateDocument functions in the module corresponding to
the absent namespace, where that namespace is used for resolution.

When the document declares a default namespace in an outer element, but
undefines it internally, PyXB did not use the module-provided fallback
namespace, instead failing to locate the information.