examples/ndfd: fix issues related to schema update
Subsequent to redirecting the service from
"http://www.weather.gov/forecasts/xml" to
"http://graphical.weather.gov/xml", the schema for NDFD were updated.
Several issues arise:

The as-distributed parameters.xsd has a pressure.units attribute for which
the default value is not one of the valid restrictions.  This requires a
manual edit of the generated bindings file to add it.  A request that this
be fixed upstream is pending on the NDFD user group at Yahoo.

(showreq.py) The WSDL now defines a Unit part that is not present in the
example document.  Update to accommodate this inconsistency.

(forecast.py) The data element changed from maxOccurs=1 to
maxOccurs=unbounded, requiring that the first element be extracted before
the contents are printed.