trac/111: Make simpleType CF_Enumeration iterable
Add values, itervalues, items, and iteritems to the enumeration_mixin class,
forwarding to the underlying CF_Enumeration instance.  The value methods
return the enumeration values, while the item methods return the
_EnumerationElement instance from which the Python value, Unicode value, and
tag identifier can be obtained.  The item methods are inherited from
_CollectionFacet_mixin.

This functionality was available in the CF_Enumeration.elements() method,
which is now deprecated in favor of items or iteritems.