trac/129: incorrect timezone adjustment from python values
Whether the instance comes from Python or XML, presence of a time zone
indicates that the described time is in that time zone, and the offset of
the timezone must be subtracted from, not added to, the time.  Use of
fromutc was completely wrong, and led to the mistaken concept of a flipped
UTC offset.