trac/131: validate processing of unicode in assignments and documents
PyXB supports three parsing mechanisms: minidom, saxdom, and saxer, with
saxer being the default.  The saxdom style is capable of processing the
unicode string without conversion.  minicom and saxer use expat underneath,
and expat only works on non-unicode strings.  In that situation, whatever's
passed to CreateFromDocument must be converted to a binary string first.

(The original report had binary data encoded in a non-unicode string created
without an encoding; per web references in the test case this is incorrect.)