Rewrite character class parsing code.
The spec says that a dash at the beginning or end of a character group
means a literal dash, it should not be a parse error.
Also fixes at least one case where the "-[" set subraction syntax was
not parsed, causing a syntax error ("[a^-[a]]")

Simplify parsing of backslash-escaped characters by having a single
dictionary keyed by every possible one.

Signed-off-by: Peter A. Bigot <pab@pabigot.com>