Change print statements to logging.
Generally, libraries shouldn't use "print".  If the library is used
by a web service or a GUI application, messages printed using "print"
will be lost.  If run as a CGI script or a console application,
messages printed with "print" will mess up the application's output.

Change to use Python's standard "logging" module instead.  This allows
the application author to decide what to do with the messages.  If
the application is already using the "logging" module, the messages
will end up in the log file automatically.

Signed-off-by: Peter A. Bigot <pab@pabigot.com>