Fix tests run with libxml2 parser
More cases where "t><n" in a document causes it to look like bogus unicode;
force the thing to utf-8.  libxml2 raises a parse exception instead of a
UnicodeEncodeError in some cases.  It also diagnoses an illegal URI.

(Note: This validated with drv_libxml2 passed as the preferred parser in
make_parser, but that's not part of the commit.)