trac/146: mark encoding in generated output
Add global values for the input and output encodings to the top-level pyxb
module, rather than hard-coding them throughout the source.

Add the output encoding used by the generation module to the top of the
generated python module.

Literal uses of specific encodings are left for pickling and for hash
calculation.  All other uses should reference the global values for now;
when trac/145 is addressed the information should derive from the encoding
of the schema or some other source.