trac/139: support alternative XML parsers
For the purpose of solving this problem, it is sufficient to use an
alternative XmlReader; it is not necessary to support XMLStyle_lxml which
would support lxml in the DOM domain.  A crude but usable interface has been
added to configure alternatives.

Note that there is a remaining issue which has been opened as trac/147.