trac/141: need regular way to replace MakeIdentifier
Provide a function to pre-process what's given to MakeIdentifier before it
strips out invalid characters and otherwise transforms the string.

The interface to this isn't ideal, and in particular there's no way to
override the standard behavior using an unmodified pyxbgen.  The solution
will be to use a modified pyxbgen specific to the environment; a crude
example is present in the test case, and a more refined one will be in the
unicode_jp example yet to be included.