Provide option to pre-load schemas from local repository
When building distributions, we start with a clean slate that does not
include the schemas for the bundles.  W3C gets progressively slower at
responding when asked to provide schemas from their site, to the extent that
repeated builds of distributions while testing fail.  Provide a back-door
that allows pre-population of the schema directory assuming a local copy is
available in a branch on the development area's git repository.