Refine solution for Unicode escape map
Create module-level maps for the category and block escapes in
pyxb.utils.unicode, to match what was done for the character escapes.  Move
the all-escape map to the XMLRE file.  Refine diagnostics to note why the
escape was invalid (in particular when a block does not exist).  Manually
verified that the diagnostics are accurate; need 2.7's assertRaisesRegexp to
do that properly.