Provide locations for element and attribute declaration and use
The declaration is the actual type data; the use is supposed to be the place
in the schema where the declaration was referenced.  Because of how
plurality data is calculated, the closest we can currently get to use is the
particle containing the element; if the element declaration in that particle
is a reference, the position of the referring declaration is not immediately
available.

Perhaps with the new content model we can replace the plurality
infrastructure.