Improve diagnostics and reduce effort converting wildcard DOM nodes
Make the use of a DOM-format instance more visible by emitting a warning,
rather than silently using it (or sometimes emitting an info).  Also stop
trying to locate and import modules in order to automatically convert
things.  Such imports never occurred in any test case.  Now that the
warnings are visible, the user should be able to ensure binding modules are
imported for namespaces that might be relevant.