trac/164: standardize element order
Instead of the total order produced by sccOrder(), which incorporates the
effect of Python hashing, use a partial order based on satisfaction of
incoming edges, then sort internally based on namespace and location of
definition.  Also need to sort the FAC elements, and the modules in an
SCC producing a group definition.