Maintain automata transitions as a list instead of a set
Where non-determinism occurs in the content model, we will want to resolve
it deterministically, preferring the accepting path that occurs first within
the schema that define the content model.  Doing this requires that the
order specified when the code to generate a binding automaton is emitted be
preserved in the automaton data structure.  Sets are subject to re-order
based on the address at which the object is placed; lists are not.