Placeholder for content model resolution
Where non-determinism is present the FAC system keeps element content
pending in closures that are not applied until the system has received
enough input to resolve the correct path.  Invoking finalizeContentModel is
a signal that no further element content will be provided through the
content automata interface.  Note that assigning content through keyword
parameters to the constructor or by direct assignment to element properties
bypasses the content model so finalization is not appropriate.