Correct test case
trac/112 improperly diagnosed a failure to create an instance of a document
where the greedy content model failed.  In fact creating the document was
not an issue; the error is in validating it, and occurred only because the
creation path threw out the structure-based information it had available to
it when validating.  In fact with both the old and new content models
validation post-DOM is fine; it's validation in absence of any structured
information about the content that fails.