Fix mistaken acceptance in multi-configuration
We don't use this class in PyXB, but it's already present, tested, and
serves as an example so it ought to work.  Configurations for sub-automata
are not in accepting state unless the superconfigurations are also
accepting.  (The mistake was fixed for PyXB in the course of validating
trac/112 but had not been backported.)