trac/189: fix misuse of StructuralBadDocumentError
One raise site to StructuralBadDocument was replaced by raise
NonPluralAppendError, as it should be a ValidationError not a
BadDocumentError.

The problematic call site is only entered when validation is disabled, so it
is appropriate to raise an exception that is not a ValidationError.
StructuralBadDocument has been modified to provide access to the container
and the value that could not be identified without validation.

The reference in the originating trac/188 to inability to invoke details()
on the exception is irrelevant as only ValidationErrors support details().