trac/190: error when exception details requires unicode
Change all __str__/str material in exceptions hierarchy to use unicode.
Where a __str__ method had been added before, have it delegate to use
unicode and convert to a string with the standard output encoding.
(Attempts to put a default delegating __str__ in the base exception class
introduced infinite recursion, so for now the delegation is done in each
class that needs it.)