trac/204: fixed elements improperly create non-empty default
Distinguish the reset value of an element from its default value, and use
its reset value when creating new instances.  This restores the behavior
prior to the fix for trac/99, so is a reversion from a "feature" of
PyXB-1.2.1, but that "feature" was a premature and incomplete solution to
trac/170 so it's best to remove it.