python3: import future print_function
This prepares for python3, reducing the number of changes from 2to3 by
defaulting to the upcoming style of print function.  As the new semantics is
supported in Python 2.6 we can do this now.

This is the unedited output of 2to3 -f print.  Only added the future import
to files that contained print statements, since the library and most of the
tests should not be using print for output, and the clue might remind me to
fix the couple cases where this expectation is violated.