trac/209: inappropriate re-processing of archived namespaces
Refactor archive/namespace management.  Remove --pre-load-archives; this
loaded components early, which implicitly enabled the namespaces to be
extended regardless of intent.  Replace this with
--import-augmentable-namespace which specifies particular namespaces that
may be extended even if content is available from archives or builtins.
Simplify the logic in import handling so only import-augmentable namespaces
cause anything to happen in an import.

Update the examples.  The --pre-load-archive directive in complex/nsext is
no longer necessary since the existing declarations are found by the attempt
to reference them.  A new complex/nsaugment test case handles the intended
situation of extending from an archive by importing an alternative schema.