cleanup: clearly distinguish text from data in xml representations
Python 3 replaces the str/unicode distinction with a data/text
distinction. This is an epiphany: PyXB has not fully embraced this
perspective up to now, contributing to past errors related to unicode.  So
here's what we have:

In Python 2, "str" is 8-bit data, whether it is encoded bytes (data) or
ASCII text.  "unicode" is text, period.  "str" objects that contain text
generally require an encoding, either inside the object or carried alongside
it.  unicode objects have no encoding.

In Python 3, "str" is text and is the same as Python 2 unicode.  "bytes" is
8-bit data, and is not substitutable for text.  Thus a "bytes" object must
somehow be associated with an encoding if it is to be interpreted as text,
and a text object must be converted to a data object by applying an
encoding.

A full XML document should have an encoding attribute in the prolog XMLDecl,
and thus should be manipulated as data.  The XML processing infrastructure
(xml.dom, saxer, lxml, whatever) operates on data, and provides text values
to the application.

Prior to this commit, PyXB would sometimes use variables like xmls to
indicate that the value was 8-bit text (str), and xmlu to indicate that the
value was unicode.  As expected of something borne of Python 2 it
mischaracterized these as conflicting text representations, and failed to
recognize any representation as data.  Python 2's StringIO module is an
anachronism that facilititates this mischaracterization, and its use
explains several anomalous unicode errors in unit tests.

Part of this commit converts all uses of StringIO to either io.StringIO
(text) or io.BytesIO (data), for clarity and future compatibility with
Python 3.

Another part of this commit carefully uses binary mode when opening files
that are expected to contain whole documents, including schema.

The rest of this commit renames some variables to be (e.g.) xmlt for a text
(unicode) representation, and xmld for a data (str or bytes) representation.

The scope of this commit is what is sufficient to get the unit tests to pass
under Python 2.  A subsequent commit will correct the unit tests to
recognize that toxml() produces data which is not comparable to text
expected output.