trac/211: invalid element with simple type misdiagnosed
Add a new exception to diagnose cases where element content (or something
that's supposed to be element content) shows up where it can't.  Use this
for both saxer and DOM-based parsing, and in a situation where an
element-bound value is used to initialize simple content.

The test cases probe the boundaries of mixed content more than other tests
have done.