cleanup: use test target name rather than searching for it
This has worked only because the last entry in the globals was the desired
one; the type is named STD_ANON not _STD_ANON, and in fact it's the element
instance quantity that we need.  This breaks under Python 3.3 which
randomizes dictionary order.  It's also unnecessary: just get the type from
the bound element.