cleanup: refactor nondeterministic test
The test is intended to verify that generating with validation disabled
produces a result even when a required element is missing.  The expected
result assumed the order of elements was also wrong.  This may be untrue for
Python 3.3 which randomizes things more than previous releases.  Update the
test to check both potential orders for contained content when not using the
content model to guide generation.