trac/91: Problem with fractionDigits violation
Change the value type for xs:decimal to be decimal.Decimal from its
former type of types.FloatType.  This should make PyXB conformant with
xs:decimal in terms of the totalDigits and fractionDigits constraining
facets.

This is an interface change: some arithmetic operations between
datatypes.decimal and Python floats are now likely to fail.