BDS refactor: move generation-related namespace maps into NamespaceContext
NamespaceContext had previously maintained a map of prefix-to-namespaces
as specified by xmlns attributes encountered when parsing.  Now the
declareNamespace function used when generating DOM instances shares the
same data structure.

The semantics of _BDSNamespaceSupport changed slightly: in the past only
some prefixes were stored in the map and all were added to the document.
Now all prefixes are stored in the map, and only ones that were retrieved
through a prefix lookup are added to the document.

This solution should also permit the same namespace to be referenced
through multiple prefixes if desired.