BDS refactor: fix long-standing "bug"
Commit 6604608083f2e3b for trac/46 from five years ago added a
positional parameter to the NamespaceContext constructor, but didn't
adjust a constructor call for the addition.  This shifted a value
intended for the recurse argument to be passed as the including_context
argument.  The error was not revealed because the default recurse value
was correct, and the control path that would have dereferenced the
invalid including_context parameter was never taken.