domutils: accept binding instances for text and attribute values
When issue #13 is fixed use of xsdLiteral() will fail to preserve the
information necessary to reconstruct all simple types: in particular
QName representations as XML text require a namespace declaration to be
in scope.  Instead of having the caller do the conversion to text, do it
down in the BDS infrastructure which knows how to handle this situation.