fix #17: review use of logging module for validation diagnostics
Eliminated one case where a log message conveyed nothing that could not
be obtained from the exception that was immediately raised.

However, there are other warnings (cases where PyXB was unable to
validate a field, or when it found multiple valid parses) that do not
raise exceptions but are still important enough that they should be
visible by default.  Those will be retained at the current warning
level.  (Still more messages are present at info level, which is not
displayed by default and will not be removed.)