fix #18: assertion failure generating extension bindings
The assertion attempted to verify that the namespace for the module from
which a declaration came was the same as the namespace for the
declaration.  In the case here that is not the case: the origin module
is pyxb.bundles.opengis.ogckml22 and the referenced declaration
namespace is http://www.w3.org/2005/Atom, which is itself imported
into ogckml22 from pyxb.bundles.opengis.atom.

However, the origin module does import the bindings for the referenced
declaration, so the subsequent code that imports it satisfies the
requirement and the assertion may be removed.

To determine the true module of origin would require recording origins
across independent binding generation sessions; this is unnecessary.

Reproducing command:

    pyxbgen \
      --archive-path='&pyxb/bundles/opengis//' \
      --schema-location=http://code.google.com/apis/kml/schema/kml22gx.xsd \
      --module=kml22gx