fix #19: failure in python deepcopy
The capabilities in Python's copy module depend on serializing and
unserializing instances to create cloned instances.  This breaks when
PyXB does not allow serialization of instances with absent namespaces.

While serialization of such instances cannot be supported in general
(because absent namespaces lack a unique identifier), it can be
supported within a PyXB invocation by generating a session-local unique
identifier as the serialized representation of an absent namespace.