fix #22: ensure bindings compatible with both Python2 and Python3
Use __future__ to treat all literals as unicode; this way we can strip
the u prefix from literals so Python 3 prior to 3.3 works.  Import six
in the bindings so we can determine whether to convert XML input to
data.

Fix issues with facet extrema and with FAC occurrence relations, both of
which tend to use Python 2 long values that, if emitted with "repr",
include a suffix L that is not acceptable to Python 3.

Generally replace use of %r formatting with %s and a delegation to
repr2to3 which implements the above transformations.