fix #27: conflict between class property and module type name
When referencing the name of a binding type defined in the current odule
from a class also defined in the module we need to account for the
possibility that the class has defined a property with the same name.
At the user level these are not in conflict (because property access is
qualified by an object reference), but during the class definition this
qualifier is absent.

I'd rather just reference sys.modules[__name__] to get the module-level
definition, but when run with "setup.py test" python generates:

    AttributeError: 'module' object has no attribute 'color_'

even though the class 'color_' is completely defined before the class
that tries to reference it through the module object.  So we create a
holder for all type bindings defined in the module.