fix #32: xsi:type missing on extended/restricted element content
All past experience with xsi:type had been with cases where the element
was declared as having an abstract type.  It's also permissible to
extend or refine the type that's expected by the schema.  When that's
happened, make sure the XML that's generated from the binding preserves
the fact-of the extension/restriction by specifying the corresponding
type.