binding/basis: restore part of xsi:type behavior
The fix for issue #32 uses xsi:type unconditionally on abstract types, a
change from the previous behavior that did so only when the value type
didn't match a superseding class associated with the abstract type.
Instead of attempting to create a situation that distinguishes the
conditions revert to how that case has been handled for the past five
years.