fix #26: validation of decimal facets
An inappropriate tearDown that changed the decimal context to a
non-default value in an unrelated test caused test-trac-0091 to truncate
results, resulting in failure to discover a too-precise result, but only
when run in the suite that first ran test-decimal.

Likely this only ever worked because in other versions test-trac-0091
was run first.