doc/extapi: remove GPL material
http://git.savannah.gnu.org/cgit/kenozooid.git/tree/doc/extapi.py was
originally adopted to link from sphinx to to epydoc-generated
documentation.  epydoc was last updated over nine years ago, so it's
dead.  A downstream packager noticed that the imported material was GPL,
being the only part of PyXB so licensed and conflicting with the
intended Apache 2.0 license.

Remove the material that came from the original file, leaving only the
ticket-related roles that were subsequently added for PyXB support.
Mark that material as Apache 2.0.

Closes #77.