structures: defer assigning element type if value validation deferred
The situation here is an element with a default value and a type that is
a complex type with (ultimately) simple content.  With the schema order
of declarations the first time the element is encountered the fact that
the complex type has simple content could not be confirmed, resulting in
a diagnostic and failure.  Instead delay the resolution of the element
declaration until the type is sufficiently resolved to confirm that the
default is compatible.

This is too complex to put together a test case.  Which means it'll
probably crop up again.

Closes #72