structures: remove mis-use of derivation method as indicating resolution
Interpreting completed resolution as having a set __derivationMethod
required that we instead store and reference __pendingDerivationMethod,
even though in fact the derivation method is fixed and was not the
characteristic that was pending.  This causes issues when it becomes
necessary to access the derivation method to control validation during
resolution.

Replace with a flag that explicitly notes resolution is complete, and
let derivation method be own true self.