builtin/XHTML: add categories to built-in
PyXB requires that http://www.w3.org/1999/xhtml be a registered
namespace as it's the default namespace for XML itself and is in scope
in other standard namespaces.  However, PyXB defines no content for it.

Through dirty tricks bindings for the modularized namespace can be
generated, but attempts to import the bindings fail if the
pre-registered namespace isn't prepared to hold the content.  Prepare
it.

See: issue #81
See: https://www.w3.org/TR/2010/REC-xhtml-modularization-20100729/