passim: fix issues with Python 3 print
When Python 3 is told to print an XML document in its encoded form it
seems to use the binary representation, which includes a prefix b and
enclosing the string in quotes.

Apparently these test cases were not in place the last time the whole
test suite was run under Python 3, because they all fail.  Rework to
ensure print is given a `str` not a `bytes`.  Also ensure print is used
as function.