Make binding generation reproducible
Make it posible to configure pyxbgen to generate bindings in a reproducible way.  "Reproducible" means that the same input results in exactly the same output, byte-for-byte, without depending on:

 * the time the bindings were built
 * the path to the working directory they're buit in
 * the MAC address of the build PC (embedded in a UUID)
 * newly generated UUIDs (guaranteed to be different every time)
 * the iteration order of Python dicts/sets
 * the path separator (e.g. "/") used by the OS that the bindings are built on.

Before this commit, the PyXB bindings changed every time they were built becasue they depended on all of the above.  After this commit, you can pass options to pyxbgen to get reproducible bindings.  The new options are:

 * --strip-file-paths - doesn't store paths to XSD files in bindings.
 * --no-timestamp - doesn't write build timestamp and Python version in comments in bindings.
 * --generation-uid= - specifies the Generation UID rather than randomly generating one.

There are two reasons why people want this:

 1) If bindings are committed to version control, then when bindings are rebuilt you want to see what actually changed and not see loads of spurious differences.

 2) If bindings are not committed to version control but are built by some build system, then careful development teams want developers to be using the exact same bindings as each other and as the build system is building.